tests/wpt/tests/css/css-scroll-snap/parsing/scroll-snap-align-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Snap: getComputedStyle().scrollSnapAlign</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#propdef-scroll-snap-align">
<meta name="assert" content="scroll-snap-align computed value is as specified.">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("scroll-snap-align", "none");
test_computed_value("scroll-snap-align", "start");
test_computed_value("scroll-snap-align", "end");
test_computed_value("scroll-snap-align", "center");

test_computed_value("scroll-snap-align", "start none");
test_computed_value("scroll-snap-align", "center end");
test_computed_value("scroll-snap-align", "start start", "start");
</script>
</body>
</html>
